requirements.txt
pydantic>=2.0
omegaconf>=2.3
numpy
pandas
matplotlib
tqdm
sentence-transformers
faiss-cpu
rank-bm25
torch
torchvision
ir-measures
google-genai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
datasets>=2.16.0
ragas>=0.4.0
langchain-openai>=1.0.0
langchain-google-genai>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0
pypdf>=4.0.0
pypdfium2>=4.30.0
rapidocr-onnxruntime>=1.4.0
minio>=7.2.0
redis>=5.0.0
pymilvus>=2.4.0
prometheus_client>=0.20.0
fastapi>=0.115.0
uvicorn>=0.30.0
